requirements.txt
requests==2.28.1
config==0.5.1
loguru==0.7.2
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0